Update transitioning_asf.ad
Edits for readability. In addition, the original text assumes a migration to a Subversion repository. My colleagues tell me that in the last few years all new podlings have imported their code into Git repositories. It might be a good idea for some wise person to review this page and confirm that it is good guidance for the current reality.